rename index.js to indexMongo.js to clarify which db it connects to
